requirements.txt
requests
pyyaml
tabulate